Use defaults to define 'modify' (save with merge) in the typeclass.
The logic for handling merges can be defined in terms of the 'retrieve'
and 'save' operations, so 'modify' is now handled by a default definition
in FileClass.  'save' is new; it is a "dumb" save-and-commit operation
that pays no attention to possible conflicts.  Implementers can define
'save' and get a sophisticated 'modify' for free.

A generic default for 'diff' is also provided.

darcs-hash:20090109080729-0f649-582410b0c490ec3de7fe5ca4fdadb9aa50d04d4b.gz