Changed isInsideDir so it doesn't use canonicalizePath.
Ignore-this: 5a908fbc8b1c2835f84a0a5310a0009c

In directory-1.1.0.0 (ghc7), canonicalizePath was changed so
that it fails if the directory does not exist.  We need this
test not to presuppose that the directory exists. So instead,
we do a simple prefix comparison, returning False if either the filename
or the directory name contains '..'.

darcs-hash:20110126162834-0f649-08c6921b56ee6b6e1c8079469be871a25ab2a168.gz